LayoutTests/imported/w3c/web-platform-tests/WebIDL/ecmascript-binding/class-string-named-properties-object.window-expected.txt


FAIL @@toStringTag exists with the appropriate descriptor assert_own_property: expected property symbol "Symbol(Symbol.toStringTag)" missing
PASS Object.prototype.toString 
FAIL Object.prototype.toString applied after modifying @@toStringTag assert_own_property: Precondition for this test: @@toStringTag exists expected property symbol "Symbol(Symbol.toStringTag)" missing
FAIL Object.prototype.toString applied after deleting @@toStringTag assert_equals: prototype expected "[object EventTarget]" but got "[object WindowProperties]"
